Refactoring: first iteration
