--- backend/helplib/__init__.py
@@ -1,14 +1,0 @@
-from helplib import (
-    checkers,
-    commands,
-    exceptions,
-    flags,
-    locking,
-    models,
-    types,
-)
-
-__all__ = [
-    'checkers', 'commands', 'exceptions', 'flags',
-    'locking', 'models', 'types',
-]
