+++ backend/lib/helpers/exceptions.py
