+++ backend/lib/models/__init__.py
@@ -1,0 +1,13 @@
+from .attack_result import AttackResult
+from .flag import Flag
+from .game_state import GameState
+from .global_config import GlobalConfig
+from .task import Task
+from .team import Team
+from .verdict import CheckerVerdict
+
+__all__ = (
+    AttackResult, Flag, GameState,
+    GlobalConfig, Task, Team,
+    CheckerVerdict,
+)
