+++ backend/lib/storage/teams.py
@@ -1,22 +1,21 @@
 from typing import List, Optional
 
-import storage
-from helplib import models
-from helplib.cache import cache_helper, async_cache_helper
-from storage import caching
+from lib import models
+from lib import storage
+from lib.helpers.cache import cache_helper, async_cache_helper
 
 
 def get_teams() -> List[models.Team]:
     """Get list of teams registered in the database."""
-    with storage.get_redis_storage().pipeline(transaction=True) as pipeline:
+    with storage.utils.get_redis_storage().pipeline(transaction=True) as pipe:
         cache_helper(
-            pipeline=pipeline,
+            pipeline=pipe,
             cache_key='teams',
-            cache_func=caching.cache_teams,
-            cache_args=(pipeline,),
+            cache_func=storage.caching.cache_teams,
+            cache_args=(pipe,),
         )
 
-        teams, = pipeline.smembers('teams').execute()
+        teams, = pipe.smembers('teams').execute()
         teams = list(models.Team.from_json(team) for team in teams)
 
     return teams
@@ -27,14 +26,14 @@ async def teams_async_getter(redis_aio, pipe) -> None:  # type: ignore
     await async_cache_helper(
         redis_aio=redis_aio,
         cache_key='teams',
-        cache_func=caching.cache_teams,
+        cache_func=storage.caching.cache_teams,
     )
     pipe.smembers('teams')
 
 
 async def get_all_teams_async() -> List[models.Team]:
     """Get list of all teams from database."""
-    async with storage.async_db_cursor(dict_cursor=True) as (_conn, curs):
+    async with storage.utils.async_db_cursor(dict_cursor=True) as (_, curs):
         await curs.execute(models.Team.get_select_all_query())
         results = await curs.fetchall()
 
@@ -49,8 +48,8 @@ def get_team_id_by_token(token: str) -> Optional[int]:
     :param token: token string
     :return: team id
     """
-    with storage.get_redis_storage().pipeline(transaction=False) as pipeline:
-        team_id, = pipeline.get(f'team:token:{token}').execute()
+    with storage.utils.get_redis_storage().pipeline(transaction=False) as pipe:
+        team_id, = pipe.get(f'team:token:{token}').execute()
 
     try:
         team_id = int(team_id)
@@ -62,7 +61,7 @@ def get_team_id_by_token(token: str) -> Optional[int]:
 
 async def create_team(team: models.Team) -> models.Team:
     """Add new team to DB, reset cache & return created instance."""
-    async with storage.async_db_cursor() as (_conn, curs):
+    async with storage.utils.async_db_cursor() as (_, curs):
         await curs.execute(team.get_insert_query(), team.to_dict())
         result, = await curs.fetchone()
         team.id = result
@@ -76,7 +75,7 @@ async def create_team(team: models.Team) -> models.Team:
         for each in insert_data:
             await curs.execute(storage.tasks.TEAMTASK_INSERT_QUERY, each)
 
-    redis_aio = await storage.get_async_redis_storage()
+    redis_aio = await storage.utils.get_async_redis_storage()
     await redis_aio.delete('teams')
 
     return team
@@ -84,10 +83,10 @@ async def create_team(team: models.Team) -> models.Team:
 
 async def update_team(team: models.Team) -> models.Team:
     """Update team, reset cache & return updated instance."""
-    async with storage.async_db_cursor() as (_conn, curs):
+    async with storage.utils.async_db_cursor() as (_, curs):
         await curs.execute(team.get_update_query(), team.to_dict())
 
-    redis_aio = await storage.get_async_redis_storage()
+    redis_aio = await storage.utils.get_async_redis_storage()
     await redis_aio.delete('teams')
 
     return team
@@ -95,8 +94,8 @@ async def update_team(team: models.Team) -> models.Team:
 
 async def delete_team(team_id: int) -> None:
     """Set active = False on a team."""
-    async with storage.async_db_cursor() as (_conn, curs):
+    async with storage.utils.async_db_cursor() as (_, curs):
         await curs.execute(models.Team.get_delete_query(), {'id': team_id})
 
-    redis_aio = await storage.get_async_redis_storage()
+    redis_aio = await storage.utils.get_async_redis_storage()
     await redis_aio.delete('teams')
