+++ backend/scripts/full_reset.py
@@ -1,14 +1,9 @@
 #!/usr/bin/env python3
 
-import sys
-from pathlib import Path
-
-BASE_DIR = Path(__file__).absolute().resolve().parents[1]
-sys.path.insert(0, str(BASE_DIR))
-
 import storage
-from scripts import reset_db, init_db
+
 from scripts import print_tokens
+from scripts import reset_db, init_db
 
 
 def run():
