+++ backend/scripts/print_tokens.py
@@ -1,19 +1,12 @@
 #!/usr/bin/env python3
 
-import sys
-
-from pathlib import Path
-
-BASE_DIR = Path(__file__).absolute().resolve().parents[1]
-sys.path.insert(0, str(BASE_DIR))
-
-import storage
+from lib import storage
 
 _SELECT_TEAMS_NAME_TOKEN_QUERY = "SELECT name, token from teams"
 
 
 def run():
-    with storage.db_cursor() as (_, curs):
+    with storage.utils.db_cursor() as (_, curs):
         curs.execute(_SELECT_TEAMS_NAME_TOKEN_QUERY)
         result = curs.fetchall()
