+++ backend/webapi/admin/base.py
@@ -1,9 +1,3 @@
-import sys
-from pathlib import Path
-
-BASE_DIR = Path(__file__).absolute().resolve().parents[1]
-sys.path.insert(0, str(BASE_DIR))
-
 from sanic import Blueprint
 from sanic.response import json as json_response
