+++ backend/webapi/admin/tasks.py
@@ -1,18 +1,9 @@
-import sys
-
-from pathlib import Path
-
-BASE_DIR = Path(__file__).absolute().resolve().parents[1]
-sys.path.insert(0, str(BASE_DIR))
-
 from sanic.response import json as json_response
 
-import storage
-from helplib import models
-
+from lib import models, storage
 from webapi.events import emit_init_scoreboard
-from .base import make_err_response
 from .api_base import ApiSet
+from .base import make_err_response
 
 
 class TaskApi(ApiSet):
